Split sidebar script into executable and lib to prepare for testing
